--- app/src/main/java/com/example/dopaminemoa/Constants.kt
@@ -1,7 +1,0 @@
-package com.example.dopaminemoa
-
-object Constants {
-
-    // 이미지 "좋아요" 상태를 저장하기 위한 Shared Preferences 키입니다.
-    var PREF_KEY = "IMAGE_LIKE_PREF"
-}
